Fixed conflicting argument - b as a buffer size
argv.b was interpreted as an option to change the name of the buffer file, but it was also used for the buffer size. Therefore, the buffer file name by default was "1.5MB". Anyway, I added an option "bp"/"path" (is that the best name for it?) that adds the ability to change buffer file path altogether (and not only the name), which is more useful.